+++ Nenis/Assets.xcassets/smile.out.symbolset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  },
+  "symbols" : [
+    {
+      "filename" : "noun_smile_symbol.svg",
+      "idiom" : "universal"
+    }
+  ]
+}
