--- Nenis/SplashViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  Nenis
-//
-//  Created by Caio Ferreira on 05/09/23.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
