+++ Nenis/features/user/UserHelper.swift
@@ -1,0 +1,8 @@
+//
+//  UserHelper.swift
+//  Nenis
+//
+//  Created by Caio Ferreira on 14/12/23.
+//
+
+import Foundation
