+++ Nenis/features/user/service/UserService.swift
@@ -1,0 +1,8 @@
+//
+//  UserService.swift
+//  Nenis
+//
+//  Created by Caio Ferreira on 11/12/23.
+//
+
+import Foundation
