+++ Nenis/features/user/ui/cell/TutorTableViewCell.swift
@@ -1,0 +1,23 @@
+//
+//  TutorTableViewCell.swift
+//  Nenis
+//
+//  Created by Caio Ferreira on 15/12/23.
+//
+
+import UIKit
+
+class TutorTableViewCell: UITableViewCell {
+
+    override func awakeFromNib() {
+        super.awakeFromNib()
+        // Initialization code
+    }
+
+    override func setSelected(_ selected: Bool, animated: Bool) {
+        super.setSelected(selected, animated: animated)
+
+        // Configure the view for the selected state
+    }
+    
+}
