+++ Nenis/remote/model/DocumentProtocol.swift
@@ -1,0 +1,12 @@
+//
+//  DocumentProtocol.swift
+//  Nenis
+//
+//  Created by Caio Ferreira on 15/12/23.
+//
+
+import Foundation
+
+protocol DocumentProtocol {
+    var id: String? { get set }
+}
